requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
pinecone
psycopg2-binary==2.9.9
PyJWT==2.8.0
google-generativeai>=0.3.0
gunicorn==20.1.0
PyMuPDF>=1.24.1
python-docx==1.1.0
langchain-text-splitters==0.3.8
tenacity>=8.1.0
werkzeug==3.0.1